feat: 新增配置文件config.ssr 分离前后端配置 close #81
