+++ example/ssr-with-dva/app/controller/page.js
@@ -1,6 +1,7 @@
 
 const Controller = require('egg').Controller
 const { renderToStream } = require('ykfe-utils')
+const ssrConfig = require('../../config/config.ssr')
 
 class PageController extends Controller {
   async index () {
@@ -9,6 +10,7 @@ class PageController extends Controller {
       // Page为webpack打包的chunkName，项目默认的entry为Page
       ctx.type = 'text/html'
       ctx.status = 200
+      Object.assign(ctx.app.config, ssrConfig)
       const stream = await renderToStream(ctx, ctx.app.config)
       ctx.body = stream
     } catch (error) {
