+++ example/ssr-with-dva/app/router.js
@@ -1,8 +1,10 @@
 'use strict'
 
+const config = require('../config/config.ssr')
+
 module.exports = app => {
   const { router, controller } = app
-  app.config.routes.map(route => {
+  config.routes.map(route => {
     router.get(`${route.path}`, controller[route.controller][route.handler])
   })
 }
