+++ example/ssr-with-dva/package.json
@@ -11,6 +11,7 @@
     "react": "16.8.3",
     "react-dev-utils": "^8.0.0",
     "react-dom": "16.8.3",
+    "react-router": "^5.0.1",
     "react-router-dom": "^5.0.0",
     "ykfe-utils": "^2.0.2"
   },
