+++ example/ssr-with-js/web/entry.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom'
 import { BrowserRouter, StaticRouter, Route } from 'react-router-dom'
 import defaultLayout from '@/layout'
 import { getWrappedComponent, getComponent } from 'ykfe-utils'
-import { routes as Routes } from '../config/config.default'
+import { routes as Routes } from '../config/config.ssr'
 
 const clientRender = async () => {
   // 客户端渲染||hydrate
