+++ example/ssr-with-mobx/web/entry.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom'
 import { BrowserRouter, StaticRouter, Route } from 'react-router-dom'
 import defaultLayout from '@/layout'
 import { getWrappedComponent, getComponent } from 'ykfe-utils'
-import { routes as Routes } from '../config/config.default'
+import { routes as Routes } from '../config/config.ssr'
 import { Provider, useStaticRendering } from 'mobx-react'
 import initStore from './store'
