+++ packages/yk-cli/package.json
@@ -1,6 +1,6 @@
 {
   "name": "yk-cli",
-  "version": "2.3.2",
+  "version": "2.4.0",
   "description": "ykcli for ssr",
   "main": "./bin/index.js",
   "types": "types/index.d.ts",
